[REFACTOR/#123] 판매 모듈 설정
