+++ feature/sell/src/main/res/layout/activity_sell_confirm.xml
