+++ feature/sell/src/main/res/layout/activity_sell_progress.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="vm"
-            type="co.orange.presentation.sell.progress.SellProgressViewModel" />
+            type="co.orange.sell.progress.SellProgressViewModel" />
 
     </data>
