+++ feature/sell/src/main/res/layout/dialog_bank.xml
